double typecast added (#3)
* Decimal typecase added

* double keyword added

+++ robocode.battle/src/main/java/net/sf/robocode/battle/peer/RobotPeer.java
@@ -682,8 +682,8 @@ public final class RobotPeer implements IRobotPeerBattle, IRobotPeer {
 			double maxWidth = battleRules.getBattlefieldWidth() - RobotPeer.WIDTH;
 			double maxHeight = battleRules.getBattlefieldHeight() - RobotPeer.HEIGHT;
 
-			double halfRobotWidth = RobotPeer.WIDTH / 2;
-			double halfRobotHeight = RobotPeer.HEIGHT / 2;
+			double halfRobotWidth = (double) RobotPeer.WIDTH / 2;
+			double halfRobotHeight = (double) RobotPeer.HEIGHT / 2;
 
 			int sentryBorderSize = battle.getBattleRules().getSentryBorderSize();
 			int sentryBorderMoveWidth = sentryBorderSize - RobotPeer.WIDTH;
